CODY-92 rename files in documentation module
